requirements.txt
argparse
scipy
opencv-python
opencv-contrib-python
matplotlib
imageio
tqdm
tensorboard
tensorboardX
tqdm
pyyaml
imageio
imgaug
jupyter
scikit-learn
torchgeometry
torchsummary
coloredlogs